Changed code of ADC. Still not working, but now less errors

+++ drivers/driverHeaders/PotInput.h
@@ -3,4 +3,4 @@
 
 
 void initPot();
-int32_t getPotInput();
+uint16_t getPotInput();
